[Breaking Change] Handle non-2xx API-only responses (#3)
This is a breaking change and will require a major version bump. Merging this into an `experimental` branch until we're ready to publicize this feature more.

This incorporates the changes from https://github.com/ChristianHansen/berbix-go/pull/3  plus some test updates in commit `34b352cdf4e4f22d1`, but with the difference that this new fork allows merging into `berbix/berbix-go`.

+++ berbix_test.go
@@ -1,6 +1,7 @@
 package berbix
 
 import (
+	"io/ioutil"
 	"os"
 	"testing"
 )
@@ -24,7 +25,7 @@ func TestCreateTransaction(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	assertTransaction(t, client, tokens)
+	assertTransaction(t, client, tokens, true)
 }
 
 func TestCreateHostedTransaction(t *testing.T) {
@@ -51,10 +52,135 @@ func TestCreateHostedTransaction(t *testing.T) {
 		t.Error("expected hosted url to be returned")
 	}
 
-	assertTransaction(t, client, &resp.Tokens)
+	assertTransaction(t, client, &resp.Tokens, true)
 }
 
-func assertTransaction(t *testing.T, client Client, tokens *Tokens) {
+func TestCreateAPIOnlyTransaction(t *testing.T) {
+	secret := os.Getenv("BERBIX_DEMO_TEST_CLIENT_SECRET")
+	host := os.Getenv("BERBIX_DEMO_API_HOST")
+	templateKey := os.Getenv("BERBIX_DEMO_API_ONLY_TEMPLATE_KEY")
+	// for simplicity, hardcode assumptions
+	const idType = "P"
+	const idCountry = "CA"
+	frontUploadPath := os.Getenv("BERBIX_SAMPLE_CA_PASSPORT_PATH")
+
+	client := NewClient(secret, &ClientOptions{
+		Host: host,
+	})
+
+	options := &CreateAPIOnlyTransactionOptions{
+		CreateTransactionOptions: CreateTransactionOptions{
+			CustomerUID: customerUID,
+			TemplateKey: templateKey,
+		},
+		APIOnlyOptions: APIOnlyOptions{
+			IDType:    idType,
+			IDCountry: idCountry,
+		},
+	}
+	createRes, err := client.CreateAPIOnlyTransaction(options)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	t.Logf("will upload image at %q", frontUploadPath)
+	frontRdr, err := os.Open(frontUploadPath)
+	if err != nil {
+		t.Fatal(err)
+	}
+	defer frontRdr.Close()
+	frontBytes, err := ioutil.ReadAll(frontRdr)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	t.Logf("total number of bytes in image: %d", len(frontBytes))
+
+	opts := &UploadImageOptions{
+		image:   frontBytes,
+		subject: ImageSubjectDocumentFront,
+		format:  ImageFormatJPEG,
+	}
+	upRes, err := client.UploadImage(&createRes.Tokens, opts)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	const expectedNextStep = "done"
+	if upRes.NextStep != expectedNextStep {
+		t.Errorf("expected next step of %q but got %q", expectedNextStep, upRes.NextStep)
+	}
+
+	opts = &UploadImageOptions{
+		image:   frontBytes,
+		subject: ImageSubjectDocumentFront,
+		format:  ImageFormatJPEG,
+	}
+	_, err = client.UploadImage(&createRes.Tokens, opts)
+	if _, ok := err.(InvalidStateErr); !ok {
+		t.Errorf("expected invalid state error, got %v", err)
+	}
+
+	// Can't override because transaction has already been completed, so just make
+	// sure we can get the transaction metadata
+	assertCustomerUIDFromAPI(t, client, &createRes.Tokens)
+}
+
+func TestOverrideAPIOnlyTransaction(t *testing.T) {
+	secret := os.Getenv("BERBIX_DEMO_TEST_CLIENT_SECRET")
+	host := os.Getenv("BERBIX_DEMO_API_HOST")
+	templateKey := os.Getenv("BERBIX_DEMO_API_ONLY_TEMPLATE_KEY")
+	// for simplicity, hardcode assumptions
+	const idType = "DL"
+	const idCountry = "US"
+
+	client := NewClient(secret, &ClientOptions{
+		Host: host,
+	})
+
+	options := &CreateAPIOnlyTransactionOptions{
+		CreateTransactionOptions: CreateTransactionOptions{
+			CustomerUID: customerUID,
+			TemplateKey: templateKey,
+		},
+		APIOnlyOptions: APIOnlyOptions{
+			IDType:    idType,
+			IDCountry: idCountry,
+		},
+	}
+	createRes, err := client.CreateAPIOnlyTransaction(options)
+	if err != nil {
+		t.Fatal(err)
+
+	}
+	frontUploadPath := os.Getenv("BERBIX_SAMPLE_DL_FRONT_PATH")
+	t.Logf("will upload image at %q", frontUploadPath)
+	frontRdr, err := os.Open(frontUploadPath)
+	if err != nil {
+		t.Fatal(err)
+	}
+	defer frontRdr.Close()
+	frontBytes, err := ioutil.ReadAll(frontRdr)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	t.Logf("total number of bytes in image: %d", len(frontBytes))
+	opts := &UploadImageOptions{
+		image:   frontBytes,
+		subject: ImageSubjectDocumentFront,
+		format:  ImageFormatJPEG,
+	}
+	upRes, err := client.UploadImage(&createRes.Tokens, opts)
+	if err != nil {
+		t.Fatal(err)
+	}
+	t.Logf("upRes: %+v", upRes)
+	const deleteTransaction = false
+	assertTransaction(t, client, &createRes.Tokens, deleteTransaction)
+}
+
+func assertTransaction(t *testing.T, client Client, tokens *Tokens, deleteTransaction bool) {
 	err := client.OverrideTransaction(tokens, &OverrideTransactionOptions{
 		ResponsePayload: "us-dl",
 		Flags: []string{
@@ -68,17 +194,10 @@ func assertTransaction(t *testing.T, client Client, tokens *Tokens) {
 		t.Fatal(err)
 	}
 
-	resultsA, err := client.FetchTransaction(tokens)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if resultsA.CustomerUID != customerUID {
-		t.Errorf("customer UID did not match expectations")
-	}
+	resultsA := assertCustomerUIDFromAPI(t, client, tokens)
 
 	if len(resultsA.Flags) != 1 {
-		t.Fatal("number of flags did not match expectations")
+		t.Fatalf("number of flags did not match expectations. Flags were: %v", resultsA.Flags)
 	}
 
 	if resultsA.Flags[0] != "id_under_21" {
@@ -86,7 +205,7 @@ func assertTransaction(t *testing.T, client Client, tokens *Tokens) {
 	}
 
 	if resultsA.Fields == nil || resultsA.Fields.GivenName == nil || resultsA.Fields.GivenName.Value != "the_name" {
-		t.Errorf("expected GivenName to be the_name but was %s", resultsA.Fields.GivenName)
+		t.Errorf("expected GivenName to be the_name but was %s", resultsA.Fields.GivenName.Value)
 	}
 
 	refreshToken := TokensFromRefresh(tokens.RefreshToken)
@@ -100,8 +219,55 @@ func assertTransaction(t *testing.T, client Client, tokens *Tokens) {
 		t.Errorf("expected matching customer UID")
 	}
 
-	if err := client.DeleteTransaction(tokens); err != nil {
+	if deleteTransaction {
+		if err := client.DeleteTransaction(tokens); err != nil {
+			t.Fatal(err)
+		}
+	}
+}
+
+func assertCustomerUIDFromAPI(t *testing.T, client Client, tokens *Tokens) *TransactionMetadata {
+	resultsA, err := client.FetchTransaction(tokens)
+	if err != nil {
 		t.Fatal(err)
+	}
+
+	if resultsA.CustomerUID != customerUID {
+		t.Errorf("customer UID did not match expectations")
+	}
+	return resultsA
+}
+
+func TestUploadOversizedImageAPIOnly(t *testing.T) {
+	secret := os.Getenv("BERBIX_DEMO_TEST_CLIENT_SECRET")
+	host := os.Getenv("BERBIX_DEMO_API_HOST")
+	templateKey := os.Getenv("BERBIX_DEMO_API_ONLY_TEMPLATE_KEY")
+
+	client := NewClient(secret, &ClientOptions{
+		Host: host,
+	})
+
+	options := &CreateAPIOnlyTransactionOptions{
+		CreateTransactionOptions: CreateTransactionOptions{
+			CustomerUID: customerUID,
+			TemplateKey: templateKey,
+		},
+		APIOnlyOptions: APIOnlyOptions{},
+	}
+	createRes, err := client.CreateAPIOnlyTransaction(options)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	tooManyBytes := make([]byte, 11*1024*1024*1024)
+	opts := &UploadImageOptions{
+		image:   tooManyBytes,
+		subject: ImageSubjectDocumentFront,
+		format:  ImageFormatJPEG,
+	}
+	_, err = client.UploadImage(&createRes.Tokens, opts)
+	if _, ok := err.(PayloadTooLargeErr); !ok {
+		t.Errorf("expected to get a PayloadTooLargeErr, but got %v", err)
 	}
 }
